Refactor move EnableAnswerOptionsEvaluator to expressions package
